modified the error handling logic and its corresponsding test cases
