+++ src/clients/AasRepositoryClient.ts
@@ -21,6 +21,7 @@ import {
     convertCoreAssetInformationToApiAssetInformation,
     convertCoreReferenceToApiReference,
 } from '../lib/convertAasTypes';
+import { handleApiError } from '../lib/errorHandler';
 
 export class AasRepositoryClient {
     /**
@@ -69,32 +70,9 @@ export class AasRepositoryClient {
                 data: { pagedResult: result.pagingMetadata, result: shells },
             };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -122,32 +100,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: convertApiAasToCoreAas(result) };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -177,32 +132,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-           // return { success: false, error: err as RequiredError };
         }
     }
 
@@ -220,7 +152,7 @@ export class AasRepositoryClient {
         aasIdentifier: string;
     }): Promise<ApiResult<AssetAdministrationShell, RequiredError>> {
         const { configuration, aasIdentifier } = options;
-        
+
         try {
             const apiInstance = new AasRepository(applyDefaults(configuration));
 
@@ -229,34 +161,12 @@ export class AasRepositoryClient {
             const result = await apiInstance.getAssetAdministrationShellById({
                 aasIdentifier: encodedAasIdentifier,
             });
-            
+
             return { success: true, data: convertApiAasToCoreAas(result) };
         } catch (err) {
-            
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
-            console.log('Error:', customError);
+            const customError = handleApiError(err);
+            //console.log('Error:', customError);
+            console.log(`Error: field: ${customError.field}, message: ${customError.message}`);
             return { success: false, error: customError };
         }
     }
@@ -290,32 +200,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-           // return { success: false, error: err as RequiredError };
         }
     }
 
@@ -348,32 +235,9 @@ export class AasRepositoryClient {
                 data: convertApiAssetInformationToCoreAssetInformation(result),
             };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -406,32 +270,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -460,32 +301,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -515,32 +333,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -576,32 +371,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -644,32 +416,9 @@ export class AasRepositoryClient {
                 },
             };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = aasIdentifier;
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -702,32 +451,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: convertApiReferenceToCoreReference(result) };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 
@@ -761,32 +487,9 @@ export class AasRepositoryClient {
 
             return { success: true, data: result };
         } catch (err) {
-            let message = "Unknown error";
-            let statusCode = "N/A";
-            let field = "";
-            
-            if (err instanceof RequiredError) {
-                field = err.field;
-                message = err.message;
-            }
-            else if (err instanceof Error) {
-                
-                message = err.toString().split('\n')[0];
-            }
-            // Try to extract HTTP status code if available
-            const response = (err as any)?.response;
-            if (response?.status) {
-                statusCode = response.status;
-                message = `HTTP ${statusCode} - ${message}`;
-            }
-            
-            const customError = new RequiredError(
-               field, 
-               message
-            );
+            const customError = handleApiError(err);
             console.log('Error:', customError);
             return { success: false, error: customError };
-            //return { success: false, error: err as RequiredError };
         }
     }
 }
